refactor app insights operations from buildpack_binding.py
